fix: don't require SECRET_KEY
In production, ideally the key should be required. But because the
settings object is also generated in development where the key is not
required, this produces a KeyError.

With an empty key, flask will raise an exception in production if a
feature requiring the secret key is used.

An alternative to this approach might be to define the environments in
separate modules and only import the one used at runtime.