chore: move travis test logic out of makefile
The makefile is generally considered a helper for often-used workflow
commands so it doesn't make sense to me to keep the travis test logic
there. Travis may use other makefile targets that are in common, but I
think it makes sense to spell out the test command (and codecov logic)
in .travis.yml.