style: ignore D100-D107 by default
Those codes[1] require all public elements to always include a
docstring. That might make sense for libraries, but I consider it too
strict for our applications. Rather than having a strict rule, it is up
to developers and reviewers to judge in what cases a docstring is
necessary.

As soon as a docstring is in place, all the formatting rules still apply
of course, as long as they are not in conflict with higher-priority
formatters (potentially black) or the numpy guide.

[1] http://www.pydocstyle.org/en/2.1.1/error_codes.html